+++ bridge-ui/src/components/layouts/Layout.tsx
@@ -17,67 +17,30 @@ export function Layout({ children }: { children: React.ReactNode }) {
   const pathname = usePathname();
 
   if (!sdkHasLoaded) {
-    return (
-      <div className="layout">
-        <div className="container-v2">
-          <Header theme={Theme.navy} />
-          <main>
-            <div className={styles["content-wrapper"]}>
-              <InternalNav />
-            </div>
-            {children}
-          </main>
-        </div>
-        <div>
-          <Image
-            className="left-illustration"
-            src={"/images/illustration/illustration-left.svg"}
-            role="presentation"
-            alt="illustration left"
-            width={300}
-            height={445}
-            priority
-          />
-          <Image
-            className="right-illustration"
-            src={"/images/illustration/illustration-right.svg"}
-            role="presentation"
-            alt="illustration right"
-            width={610}
-            height={842}
-            priority
-          />
-          <Image
-            className={clsx("mobile-illustration", { hidden: pathname === "/faq" })}
-            src={"/images/illustration/illustration-mobile.svg"}
-            role="presentation"
-            alt="illustration mobile"
-            width={0}
-            height={0}
-            style={{ width: "100%", height: "auto", objectFit: "cover" }}
-            priority
-          />
-        </div>
-      </div>
-    );
+    return <CommonLayout pathname={pathname}>{children}</CommonLayout>;
   }
 
+  return <CommonLayout pathname={pathname}>{children}</CommonLayout>;
+}
+
+function CommonLayout({ children, pathname }: { children: React.ReactNode; pathname: string }) {
   return (
     <div className="layout">
       <div className="container-v2">
         <Header theme={Theme.navy} />
         <main>
-          <div className={styles["content-wrapper"]}>
-            <InternalNav />
-          </div>
+          {pathname !== "/faq" && (
+            <div className={styles["content-wrapper"]}>
+              <InternalNav hide={pathname === "/"} />
+            </div>
+          )}
           {children}
         </main>
       </div>
-
       <div>
         <Image
           className="left-illustration"
-          src={"/images/illustration/illustration-left.svg"}
+          src="/images/illustration/illustration-left.svg"
           role="presentation"
           alt="illustration left"
           width={300}
@@ -86,7 +49,7 @@ export function Layout({ children }: { children: React.ReactNode }) {
         />
         <Image
           className="right-illustration"
-          src={"/images/illustration/illustration-right.svg"}
+          src="/images/illustration/illustration-right.svg"
           role="presentation"
           alt="illustration right"
           width={610}
@@ -95,7 +58,7 @@ export function Layout({ children }: { children: React.ReactNode }) {
         />
         <Image
           className={clsx("mobile-illustration", { hidden: pathname === "/faq" })}
-          src={"/images/illustration/illustration-mobile.svg"}
+          src="/images/illustration/illustration-mobile.svg"
           role="presentation"
           alt="illustration mobile"
           width={0}