+++ bridge-ui/src/components/modal/index.tsx
@@ -1,8 +1,8 @@
 import { createPortal } from "react-dom";
-import { motion, AnimatePresence } from "framer-motion";
+import { motion, AnimatePresence } from "motion/react";
 import CloseIcon from "@/assets/icons/close.svg";
 import styles from "./modal.module.scss";
-import { useEffect, useState } from "react";
+import { JSX, useEffect, useState } from "react";
 import clsx from "clsx";
 
 type Props = {